Simplify input region handling
The qml side created a Qt.rect() instance, then split it into
coordinate and size components to be passed to c++, which then
used the properties to construct a QRect. But then again did that only
for one of the three masks it support, others handled directly as rects.

It doesn't really matter that much what gets set into specific properties,
but let's just call this overlayMask as that's what it's used for.
The other two have already target specific naming.
