+++ apps/captiveportal/qml/pages/CaptivePortalPage.qml
@@ -21,7 +21,6 @@ import "../shared" as Shared
 Page {
     id: browserPage
 
-    readonly property rect inputMask: inputMaskForOrientation(orientation)
     readonly property bool active: status == PageStatus.Active
 
     property alias overlay: overlay
@@ -170,10 +169,7 @@ Page {
 
     InputRegion {
         window: webView.chromeWindow
-        x: inputMask.x
-        y: inputMask.y
-        width: inputMask.width
-        height: inputMask.height
+        overlayMask: inputMaskForOrientation(orientation)
     }
 
     Browser.CaptivePortalOverlay {
