+++ apps/core/inputregion.h
@@ -22,29 +22,16 @@ class InputRegionPrivate;
 class InputRegion : public QObject
 {
     Q_OBJECT
-// add x, y
-    Q_PROPERTY(qreal x READ x WRITE setX NOTIFY xChanged FINAL)
-    Q_PROPERTY(qreal y READ y WRITE setY NOTIFY yChanged FINAL)
-    Q_PROPERTY(qreal width READ width WRITE setWidth NOTIFY widthChanged FINAL)
-    Q_PROPERTY(qreal height READ height WRITE setHeight NOTIFY heightChanged FINAL)
+    Q_PROPERTY(QRect overlayMask READ overlayMask WRITE setOverlayMask NOTIFY overlayMaskChanged)
     Q_PROPERTY(QRect selectionStartHandleMask READ selectionStartHandleMask WRITE setSelectionStartHandleMask NOTIFY selectionStartHandleMaskChanged FINAL)
     Q_PROPERTY(QRect selectionEndHandleMask READ selectionEndHandleMask WRITE setSelectionEndHandleMask NOTIFY selectionEndHandleMaskChanged FINAL)
     Q_PROPERTY(QWindow *window READ window WRITE setWindow NOTIFY windowChanged FINAL)
+
 public:
     InputRegion(QObject *parent = 0);
 
-    qreal x() const;
-    void setX(qreal x);
-
-    qreal y() const;
-    void setY(qreal y);
-
-    qreal width() const;
-    void setWidth(qreal width);
-
-    qreal height() const;
-    void setHeight(qreal height);
-
+    const QRect& overlayMask() const;
+    void setOverlayMask(const QRect& rect);
     const QRect& selectionStartHandleMask() const;
     void setSelectionStartHandleMask(const QRect& rect);
     const QRect& selectionEndHandleMask() const;
@@ -54,10 +41,7 @@ public:
     void setWindow(QWindow *window);
 
 signals:
-    void xChanged();
-    void yChanged();
-    void widthChanged();
-    void heightChanged();
+    void overlayMaskChanged();
     void selectionStartHandleMaskChanged();
     void selectionEndHandleMaskChanged();
     void windowChanged();
